Add async example and refactor tests for aliyun_oss_x

+++ tests/test_bucket_style.py
@@ -1,46 +1,61 @@
-from .common import *
+import unittest
+
+import aliyun_oss_x
+
+from .common import OssTestCase
+
 
 class TestBucketStyle(OssTestCase):
     def test_bucket_style_normal(self):
-        result = self.bucket.put_bucket_style('imagestyle','image/resize,w_200')
+        if self.bucket is None:
+            raise Exception("bucket is None")
+
+        result = self.bucket.put_bucket_style("imagestyle", "image/resize,w_200")
         self.assertEqual(200, result.status)
 
-        get_result = self.bucket.get_bucket_style('imagestyle')
-        self.assertEqual('imagestyle', get_result.name)
-        self.assertEqual('image/resize,w_200', get_result.content)
+        get_result = self.bucket.get_bucket_style("imagestyle")
+        self.assertEqual("imagestyle", get_result.name)
+        self.assertEqual("image/resize,w_200", get_result.content)
 
-        result = self.bucket.put_bucket_style('imagestyle1','image/resize,w_300')
+        result = self.bucket.put_bucket_style("imagestyle1", "image/resize,w_300")
         self.assertEqual(200, result.status)
 
         list_result = self.bucket.list_bucket_style()
-        self.assertEqual('imagestyle', list_result.styles[0].name)
-        self.assertEqual('image/resize,w_200', list_result.styles[0].content)
-        self.assertEqual('imagestyle1', list_result.styles[1].name)
-        self.assertEqual('image/resize,w_300', list_result.styles[1].content)
+        self.assertEqual("imagestyle", list_result.styles[0].name)
+        self.assertEqual("image/resize,w_200", list_result.styles[0].content)
+        self.assertEqual("imagestyle1", list_result.styles[1].name)
+        self.assertEqual("image/resize,w_300", list_result.styles[1].content)
 
-        del_result = self.bucket.delete_bucket_style('imagestyle')
+        del_result = self.bucket.delete_bucket_style("imagestyle")
         self.assertEqual(204, del_result.status)
 
-        del2_result = self.bucket.delete_bucket_style('imagestyle1')
+        del2_result = self.bucket.delete_bucket_style("imagestyle1")
         self.assertEqual(204, del2_result.status)
 
     def test_bucket_style_exception(self):
+        if self.bucket is None:
+            raise Exception("bucket is None")
         try:
-            self.bucket.put_bucket_style('imagestyle','')
-        except oss2.exceptions.ServerError as e:
-            self.assertEqual(e.details['Message'], 'Style content is invalid.')
+            self.bucket.put_bucket_style("imagestyle", "")
+        except aliyun_oss_x.exceptions.ServerError as e:
+            self.assertEqual(e.details["Message"], "Style content is invalid.")
 
     def test_bucket_style_exception_style_name(self):
+        if self.bucket is None:
+            raise Exception("bucket is None")
         try:
-            self.bucket.put_bucket_style('@%$^&%(&(*(*','image/resize,w_200')
-        except oss2.exceptions.ServerError as e:
-            self.assertEqual(e.details['Code'], 'InvalidArgument')
+            self.bucket.put_bucket_style("@%$^&%(&(*(*", "image/resize,w_200")
+        except aliyun_oss_x.exceptions.ServerError as e:
+            self.assertEqual(e.details["Code"], "InvalidArgument")
 
     def test_bucket_style_content(self):
-        self.bucket.put_bucket_style('aaa','csdar5324')
-        get_result = self.bucket.get_bucket_style('aaa')
-        self.assertEqual('aaa', get_result.name)
-        self.assertEqual('csdar5324', get_result.content)
+        if self.bucket is None:
+            raise Exception("bucket is None")
+        self.bucket.put_bucket_style("aaa", "csdar5324")
+        get_result = self.bucket.get_bucket_style("aaa")
+        self.assertEqual("aaa", get_result.name)
+        self.assertEqual("csdar5324", get_result.content)
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     unittest.main()
